move ClustersWorkBase to moe logical place
